api/requirements.txt
wheel
requests

uvicorn
websockets
fastapi
fastapi-cache2[redis]
fastapi-cloudauth
